Move __capability to pointer type if specified before. Output warning/errors.
Clang will now treat a __capability qualifier written before the pointee type
to be a qualifier on the pointer type (e.g. __capability int *). In C99,
pointer type qualifiers should be after the *, hence we output a deprecated
usage warning together with a FixItHint that will automatically transform the
source code to put it in the right place if you specify -Xclang -fixit.

Currently, the only time you will not get a warning is if the pointee type is a
typedef of a pointer type, because then it is actually the pointer type that is
being qualified. This is consistent with const/volatile in C99. For example:

typedef int * intptr;
__capability intptr x;

Here, x is a capability-qualified pointer to int.

This handling of __capability also works for function pointer types and array
types. For example, it will give the correct semantics for these (albeit with a
warning):

__capability int (*f)(void);
__capability int (*p)[5];

Clang will also output an error if the deprecated use of __capability is ambiguous. For example, the following will generate an error:

__capability int ** x;

Finally, Clang will output an error if __capability has not been applied to a
pointer type, such as in:

__capability int x;